requirements.txt
absl-py==2.1.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
archspec @ file:///croot/archspec_1709217642129/work
asttokens==2.4.1
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///work/perseverance-python-buildout/croot/brotli-split_1698805593785/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1718025014955/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.89
cloudpickle==3.0.0
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1715631928597/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1698882886606/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///work/perseverance-python-buildout/croot/conda-package-handling_1698851267218/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
contourpy==1.3.0
cryptography @ file:///croot/cryptography_1710350347627/work
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
distro @ file:///work/perseverance-python-buildout/croot/distro_1701732366176/work
dm-tree==0.1.8
einops==0.8.0
etils==1.12.0
executing==2.1.0
Farama-Notifications==0.0.4
fastprogress==1.0.3
filelock==3.16.1
fire==0.7.0
flax==0.10.4
fonttools==4.53.1
frozendict @ file:///home/conda/feedstock_root/build_artifacts/frozendict_1715092752354/work
fsspec==2025.2.0
gast==0.6.0
gymnasium==0.29.1
humanize==4.12.1
idna @ file:///work/perseverance-python-buildout/croot/idna_1698845632828/work
importlib_resources==6.5.2
ipython==8.28.0
jax==0.5.1
jax-cuda12-pjrt==0.5.0
jax-cuda12-plugin==0.5.0
jaxlib==0.5.1
jaxtyping==0.2.38
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1710807507480/work
jsonpointer==2.1
kiwisolver==1.4.7
libmambapy @ file:///croot/mamba-split_1712091911343/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1706732933928/work
ml_collections==1.0.0
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.8.61
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.6
osfclient==0.0.5
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
prompt_toolkit==3.0.48
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat @ file:///work/perseverance-python-buildout/croot/pycosat_1698863456259/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.19.1
pyparsing==3.1.4
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests @ file:///croot/requests_1707355572290/work
rich==13.9.4
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
sbx-rl @ git+https://github.com/araffin/sbx@19c85a1fd8fd56437c2f9c1de081889e49554604
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
simplejson==3.20.1
six==1.16.0
stable_baselines3==2.4.0a7
stack-data==0.6.3
sympy==1.13.2
tensorflow-probability==0.24.0
tensorstore==0.1.72
termcolor==2.5.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.4.1
tqdm @ file:///work/perseverance-python-buildout/croot/tqdm_1701735729845/work
traitlets==5.14.3
treescope==0.1.9
triton==3.0.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1707770551213/work
wadler_lindig==0.1.3
wcwidth==0.2.13
wheel==0.41.2
zipp==3.21.0
zstandard @ file:///work/perseverance-python-buildout/croot/zstandard_1698847073368/work
